Revert "CYS - Move the `private/patterns` endpoint to woocommerce admin API (#50400)"
This reverts commit 091141218ed0f27e717f479f4bac52e8b594e5a9.
