+++ plugins/woocommerce-admin/client/customize-store/assembler-hub/sidebar/sidebar-navigation-screen-homepage-ptk/sidebar-navigation-screen-homepage-ptk.tsx
@@ -264,7 +264,7 @@ export const SidebarNavigationScreenHomepagePTK = ( {
 											<Button
 												onClick={ async () => {
 													await apiFetch( {
-														path: `/wc-admin/patterns`,
+														path: `/wc/private/patterns`,
 														method: 'POST',
 													} );
