+++ plugins/woocommerce-admin/client/customize-store/intro/services.ts
@@ -106,7 +106,7 @@ const fetchIsFontLibraryAvailable = async () => {
 const fetchIsPTKPatternsAPIAvailable = async () => {
 	try {
 		await apiFetch( {
-			path: '/wc-admin/patterns',
+			path: '/wc/private/patterns',
 			method: 'GET',
 		} );
