+++ plugins/woocommerce/src/Admin/API/Init.php
@@ -93,7 +93,6 @@ class Init {
 				'Automattic\WooCommerce\Admin\API\AI\Patterns',
 				'Automattic\WooCommerce\Admin\API\AI\Product',
 				'Automattic\WooCommerce\Admin\API\AI\Products',
-				'Automattic\WooCommerce\Admin\API\Patterns',
 			);
 		}
