+++ plugins/woocommerce/tests/e2e-pw/tests/customize-store/assembler/homepage.spec.js
@@ -259,10 +259,9 @@ test.describe( 'Homepage tracking banner', () => {
 				'no'
 			);
 
-			await page.route( '**/wp-json/wc-admin/patterns*', ( route ) => {
-				route.fulfill( {
-					status: 500,
-				} );
+		await page.route( '**/wp-json/wc/private/patterns*', ( route ) => {
+			route.fulfill( {
+				status: 500,
 			} );
 
 			await prepareAssembler( pageObject, baseURL );
